`.github/workflows/pull-request.yaml`: Rework `e2e` job to use `kind`
Replace `e2e` job with a job that runs on a `kind` cluster.

Remove `deploy` step as it is no longer needed, i.e., we are building the images locally in the CI.

Remove `cleanup` step as it is no longer needed, i.e., we are not deploying on the GKE cluster anymore.

Remove `build-manager-docker` as it is no longer needed, i.e., we are not deploying on the GKE cluster anymore.

---

Signed-off-by: Mohamed Bana <mohamed@bana.io>